WIP refactoring TestScope code into the helpers
This code is copied and pasted so we need to have it made more common
