+++ tests/emails_s3_tests.rs
@@ -3,30 +3,18 @@ mod helpers;
 
 use std::collections::HashMap;
 use std::env;
-use std::sync::Arc;
 use std::thread;
 use std::time::Duration;
 
-use tokio::runtime::Runtime;
-
 use chrono::prelude::*;
-use serde_json::json;
 use serial_test::serial;
-use tokio_util::sync::CancellationToken;
 use uuid::Uuid;
 
 use kafka_delta_ingest::{start_ingest, IngestOptions};
 use rusoto_core::Region;
 use rusoto_s3::{CopyObjectRequest, S3};
-use tokio::task::JoinHandle;
 
-const TEST_APP_ID: &str = "emails_test";
-const TEST_CONSUMER_GROUP_ID: &str = "kafka_delta_ingest_emails";
-const TEST_PARTITIONS: i32 = 4;
-const TEST_TOTAL_MESSAGES: i32 = 200;
-
-const WORKER_1: &str = "WORKER-1";
-const WORKER_2: &str = "WORKER-2";
+use helpers::*;
 
 #[tokio::test(flavor = "multi_thread")]
 #[serial]
@@ -40,16 +28,12 @@ async fn when_rebalance_happens() {
     run_emails_s3_tests(true).await;
 }
 
-struct TestScope {
-    pub topic: String,
-    pub table: String,
-    pub workers_token: Arc<CancellationToken>,
-    pub runtime: HashMap<&'static str, Runtime>,
-}
-
 async fn run_emails_s3_tests(initiate_rebalance: bool) {
     helpers::init_logger();
-    let scope = TestScope::new().await;
+    let topic = format!("emails_s3-{}", Uuid::new_v4());
+    let table = prepare_table(&topic).await;
+    let options = create_options(helpers::WORKER_1);
+    let scope = TestScope::new(&topic, &table, options).await;
 
     let w1 = scope.create_and_start(WORKER_1).await;
 
@@ -89,138 +73,40 @@ async fn run_emails_s3_tests(initiate_rebalance: bool) {
     scope.shutdown();
 }
 
-impl TestScope {
-    async fn new() -> Self {
-        let topic = format!("emails-{}", Uuid::new_v4());
-        let table = prepare_table(&topic).await;
-        let workers_token = Arc::new(CancellationToken::new());
-        let mut runtime = HashMap::new();
-        runtime.insert(WORKER_1, helpers::create_runtime(WORKER_1));
-        runtime.insert(WORKER_2, helpers::create_runtime(WORKER_2));
+fn create_options(name: &str) -> IngestOptions {
+    env::set_var("AWS_S3_LOCKING_PROVIDER", "dynamodb");
+    env::set_var("AWS_REGION", "us-east-2");
+    env::set_var("AWS_STORAGE_ALLOW_HTTP", "true");
+    env::set_var("DYNAMO_LOCK_TABLE_NAME", "locks");
+    env::set_var("DYNAMO_LOCK_OWNER_NAME", name);
+    env::set_var("DYNAMO_LOCK_PARTITION_KEY_VALUE", "emails_s3_tests");
+    env::set_var("DYNAMO_LOCK_REFRESH_PERIOD_MILLIS", "100");
+    env::set_var("DYNAMO_LOCK_ADDITIONAL_TIME_TO_WAIT_MILLIS", "100");
+    env::set_var("DYNAMO_LOCK_LEASE_DURATION", "2");
 
-        println!("Topic: {}", &topic);
-        println!("Table: {}", &table);
-        helpers::create_topic(topic.as_str(), TEST_PARTITIONS).await;
+    let mut additional_kafka_settings = HashMap::new();
+    additional_kafka_settings.insert("auto.offset.reset".to_string(), "earliest".to_string());
+    let additional_kafka_settings = Some(additional_kafka_settings);
 
-        Self {
-            topic,
-            table,
-            workers_token,
-            runtime,
-        }
-    }
+    let allowed_latency = 2;
+    let max_messages_per_batch = 10;
+    let min_bytes_per_file = 370;
 
-    fn shutdown(self) {
-        for (_, rt) in self.runtime {
-            rt.shutdown_background()
-        }
-    }
+    let mut transforms = HashMap::new();
+    transforms.insert("date".to_string(), "substr(timestamp,`0`,`10`)".to_string());
+    transforms.insert("_kafka_offset".to_string(), "kafka.offset".to_string());
 
-    async fn create_and_start(&self, name: &str) -> JoinHandle<()> {
-        let rt = self.runtime.get(name).unwrap();
-        let topic = self.topic.clone();
-        let table = self.table.clone();
-        let options = self.create_options(name);
-        let token = self.workers_token.clone();
-        rt.spawn(async move {
-            let res = start_ingest(topic, table, options, token.clone()).await;
-            res.unwrap_or_else(|e| println!("AN ERROR OCCURED: {}", e));
-            println!("Ingest process exited");
-
-            token.cancel();
-        })
-    }
-
-    fn create_options(&self, name: &str) -> IngestOptions {
-        env::set_var("AWS_S3_LOCKING_PROVIDER", "dynamodb");
-        env::set_var("AWS_REGION", "us-east-2");
-        env::set_var("AWS_STORAGE_ALLOW_HTTP", "true");
-        env::set_var("DYNAMO_LOCK_TABLE_NAME", "locks");
-        env::set_var("DYNAMO_LOCK_OWNER_NAME", name);
-        env::set_var("DYNAMO_LOCK_PARTITION_KEY_VALUE", "emails_s3_tests");
-        env::set_var("DYNAMO_LOCK_REFRESH_PERIOD_MILLIS", "100");
-        env::set_var("DYNAMO_LOCK_ADDITIONAL_TIME_TO_WAIT_MILLIS", "100");
-        env::set_var("DYNAMO_LOCK_LEASE_DURATION", "2");
-
-        let mut additional_kafka_settings = HashMap::new();
-        additional_kafka_settings.insert("auto.offset.reset".to_string(), "earliest".to_string());
-        let additional_kafka_settings = Some(additional_kafka_settings);
-
-        let allowed_latency = 2;
-        let max_messages_per_batch = 10;
-        let min_bytes_per_file = 370;
-
-        let mut transforms = HashMap::new();
-        transforms.insert("date".to_string(), "substr(timestamp,`0`,`10`)".to_string());
-        transforms.insert("_kafka_offset".to_string(), "kafka.offset".to_string());
-
-        IngestOptions {
-            transforms,
-            kafka_brokers: helpers::test_broker(),
-            consumer_group_id: TEST_CONSUMER_GROUP_ID.to_string(),
-            app_id: TEST_APP_ID.to_string(),
-            additional_kafka_settings,
-            allowed_latency,
-            max_messages_per_batch,
-            min_bytes_per_file,
-            write_checkpoints: true,
-            ..Default::default()
-        }
-    }
-
-    async fn send_messages(&self, amount: i32) {
-        let producer = helpers::create_producer();
-        let now: DateTime<Utc> = Utc::now();
-
-        println!("Sending {} messages to {}", amount, &self.topic);
-        for n in 0..amount {
-            let json = &json!({
-                "id": n.to_string(),
-                "sender": format!("sender-{}@example.com", n),
-                "recipient": format!("recipient-{}@example.com", n),
-                "timestamp": (now + chrono::Duration::seconds(1)).to_rfc3339_opts(SecondsFormat::Secs, true),
-            });
-            helpers::send_json(&producer, &self.topic, &json).await;
-        }
-        println!("All messages are sent");
-    }
-
-    async fn wait_on_total_offset(&self, apps: Vec<String>, offset: i32) {
-        let mut table = deltalake::open_table(&self.table).await.unwrap();
-        let expected_total = offset - TEST_PARTITIONS;
-        loop {
-            table.update().await.unwrap();
-            let mut total = 0;
-            for key in apps.iter() {
-                total += table
-                    .get_app_transaction_version()
-                    .get(key)
-                    .map(|x| *x)
-                    .unwrap_or(0);
-            }
-
-            if total >= expected_total as i64 {
-                self.workers_token.cancel();
-                println!("All messages are in delta");
-                return;
-            }
-
-            println!("Expecting offsets in delta {}/{}...", total, expected_total);
-            tokio::time::sleep(Duration::from_secs(1)).await;
-        }
-    }
-
-    async fn validate_data(&self) {
-        let table = deltalake::open_table(&self.table).await.unwrap();
-        let result = helpers::read_files_from_store(&table).await;
-        let r: Vec<i32> = (0..TEST_TOTAL_MESSAGES).collect();
-        println!("Got messages {}/{}", result.len(), TEST_TOTAL_MESSAGES);
-
-        if result.len() != TEST_TOTAL_MESSAGES as usize {
-            helpers::inspect_table(&self.table).await;
-        }
-
-        assert_eq!(result, r);
+    IngestOptions {
+        transforms,
+        kafka_brokers: helpers::test_broker(),
+        consumer_group_id: helpers::TEST_CONSUMER_GROUP_ID.to_string(),
+        app_id: helpers::TEST_APP_ID.to_string(),
+        additional_kafka_settings,
+        allowed_latency,
+        max_messages_per_batch,
+        min_bytes_per_file,
+        write_checkpoints: true,
+        ..Default::default()
     }
 }
